MINOR: Move Murmur3 to Streams
