+++ streams/src/test/java/org/apache/kafka/streams/state/internals/Murmur3Test.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.utils;
+package org.apache.kafka.streams.state.internals;
 
 import static org.junit.Assert.*;
